Working server

+++ src/pl/zawody/Beans/BiegEJB.java
@@ -42,7 +42,7 @@ public class BiegEJB {
         }
     }
 
-    public Biegi find(Long id) {
+    public Biegi find(int id) {
         return entityManager.find(Biegi.class, id);
     }
